requirements.txt
Django>=4.2
djangorestframework
djangorestframework-simplejwt
psycopg2-binary
pytest
pytest-django
pytest-cov
coverage
gunicorn
python-dotenv
flake8
